+++ cmd/testclient/skill/skill_cannon.go
@@ -48,8 +48,8 @@ func (p *cannon) Process() (bool, error) {
 				TTL:           1,
 				TargetType:    damage.TargetOtherClient,
 				HitEffectType: effect.TypeCannonHitEffect,
-				ViewOfsX:      int(rand.Intn(2*5) - 5),
-				ViewOfsY:      int(rand.Intn(2*5) - 5),
+				ViewOfsX:      rand.Intn(2*5) - 5,
+				ViewOfsY:      rand.Intn(2*5) - 5,
 				BigDamage:     true,
 			})
 		}
